requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
bcrypt==4.2.0
cffi==1.17.1
click==8.1.7
cryptography==44.0.0
fastapi==0.115.5
h11==0.14.0
idna==3.10
Jinja2==3.1.4
jose==1.0.0
MarkupSafe==3.0.2
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.10.1
python-multipart==0.0.17
sniffio==1.3.1
starlette==0.41.2
typing_extensions==4.12.2
uvicorn==0.32.0
